classes in modules.py
got rid of "Distorted" class
